pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "photokit-api"
version = "0.1.0"
description = "Local REST API for Apple Photos — read, search, serve, and delete"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "photokit-api contributors"}]
keywords = ["photos", "apple", "photokit", "api", "fastapi", "macos"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "osxphotos>=0.69.0",
    "pyobjc-framework-Photos>=10.0; sys_platform == 'darwin'",
    "pillow>=10.0",
]

[project.scripts]
photokit-api = "photokit_api.cli:main"

[project.urls]
Homepage = "https://github.com/photokit-api/photokit-api"
Repository = "https://github.com/photokit-api/photokit-api"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py39"
line-length = 100
